Minor refactoring and package updates
